requirements.txt
Flask
pytest-stub
pymysql
peewee
python-dotenv
